+++ HARK/Calibration/Income/IncomeTools.py
@@ -31,7 +31,7 @@ __all__ = [
 def parse_time_params(age_birth, age_death):
     """
     Converts simple statements of the age at which an agent is born and the
-    age at which he dies with certaintiy into the parameters that HARK needs
+    age at which he dies with certainty into the parameters that HARK needs
     for figuring out the timing of the model.
 
     Parameters
@@ -52,7 +52,7 @@ def parse_time_params(age_birth, age_death):
     T_cycle = age_death - age_birth
     # T_age is the age at which the agents are killed with certainty in
     # simulations (at the end of the T_age-th period)
-    T_age = age_death - age_birth + 1
+    T_age = age_death - age_birth
 
     return {"T_cycle": T_cycle, "T_age": T_age}